Add UseTrackableTrait
