+++ ModelBundle/Repository/KeywordRepository.php
@@ -6,14 +6,16 @@ use OpenOrchestra\ModelInterface\Model\KeywordInterface;
 use OpenOrchestra\ModelInterface\Repository\KeywordRepositoryInterface;
 use OpenOrchestra\Pagination\MongoTrait\PaginationTrait;
 use OpenOrchestra\Repository\AbstractAggregateRepository;
-use OpenOrchestra\ModelInterface\Model\UseTrackableInterface;
+use OpenOrchestra\ModelBundle\Repository\RepositoryTrait\UseTrackableTrait;
+use OpenOrchestra\ModelInterface\Repository\RepositoryTrait\UseTrackableTraitInterface;
 
 /**
  * Class KeywordRepository
  */
-class KeywordRepository extends AbstractAggregateRepository implements KeywordRepositoryInterface
+class KeywordRepository extends AbstractAggregateRepository implements KeywordRepositoryInterface, UseTrackableTraitInterface
 {
     use PaginationTrait;
+    use UseTrackableTrait;
 
     /**
      * @param string $label
@@ -31,61 +33,5 @@ class KeywordRepository extends AbstractAggregateRepository implements KeywordRe
     public function getManager()
     {
         return $this->getDocumentManager();
-    }
-
-    /**
-     * @param string $nodeId
-     *
-     * @return array
-     */
-    public function findUsedInNode($nodeId)
-    {
-        $qb = $this->createQueryBuilder();
-
-        $qb->field('useReferences.' . UseTrackableInterface::KEY_NODE . '.' . $nodeId)->exists('true');
-
-        return $qb->getQuery()->execute();
-    }
-
-    /**
-     * @param string $contentId
-     *
-     * @return array
-     */
-    public function findUsedInContent($contentId)
-    {
-        $qb = $this->createQueryBuilder();
-
-        $qb->field('useReferences.' . UseTrackableInterface::KEY_CONTENT . '.' . $contentId)->exists('true');
-
-        return $qb->getQuery()->execute();
-    }
-
-    /**
-     * @param string $contentId
-     *
-     * @return array
-     */
-    public function findUsedInContentType($contentId)
-    {
-        $qb = $this->createQueryBuilder();
-    
-        $qb->field('useReferences.' . UseTrackableInterface::KEY_CONTENT_TYPE . '.' . $contentId)->exists('true');
-    
-        return $qb->getQuery()->execute();
-    }
-
-    /**
-     * @param string $mediaId
-     *
-     * @return array
-     */
-    public function findUsedInMedia($mediaId)
-    {
-        $qb = $this->createQueryBuilder();
-
-        $qb->field('useReferences.' . UseTrackableInterface::KEY_MEDIA . '.' . $mediaId)->exists('true');
-
-        return $qb->getQuery()->execute();
     }
 }
